+++ test/models/test_create_blob_request.py
@@ -18,47 +18,36 @@ import pydantic_core
 from datetime import datetime
 import freeclimb
 from freeclimb import *
-from freeclimb.models.say import Say
+from freeclimb.models.create_blob_request import CreateBlobRequest
 
 
-class TestSay(unittest.TestCase):
-    """Say unit test stubs"""
+class TestCreateBlobRequest(unittest.TestCase):
+    """CreateBlobRequest unit test stubs"""
 
     def setUp(self):
-        self.model = Say(
-            text="TS",
+        self.model = CreateBlobRequest(
+            blob={},
         )
 
-    def test_text(self):
-        """Test Say.text"""
+    def test_alias(self):
+        """Test CreateBlobRequest.alias"""
+        self.model.alias = "T" * 64
+        assert self.model.alias == "T" * 64
 
-        self.model.text = "TEST_STRING"
-        assert self.model.text == "TEST_STRING"
+        with self.assertRaises(pydantic_core._pydantic_core.ValidationError) as info:
+            self.model.alias = "T" * (64 + 1)
 
-    def test_language(self):
-        """Test Say.language"""
+    def test_expires_at(self):
+        """Test CreateBlobRequest.expires_at"""
 
-        self.model.language = "TEST_STRING"
-        assert self.model.language == "TEST_STRING"
+        self.model.expires_at = "TEST_STRING"
+        assert self.model.expires_at == "TEST_STRING"
 
-    def test_engine(self):
-        """Test Say.engine"""
-        object = freeclimb.models.tts_engine.TTSEngine(
-            name="freeclimb.standard",
-            parameters={},
-        )
-        self.model.engine = object
-        assert self.model.engine == object
-
-    def test_loop(self):
-        """Test Say.loop"""
-        self.model.loop = 1
-        assert self.model.loop == 1
-
-    def test_privacy_mode(self):
-        """Test Say.privacy_mode"""
-        self.model.privacy_mode = False
-        assert self.model.privacy_mode == False
+    def test_blob(self):
+        """Test CreateBlobRequest.blob"""
+        object = {}
+        self.model.blob = object
+        assert self.model.blob == object
 
 
 if __name__ == "__main__":